Split execution source context tests
