+++ tests/unit/dpm/construction/test_source_product_context.py
@@ -5,9 +5,6 @@ from src.api.services.construction_client_profile_source_context import (
     client_restriction_profile_context,
     sustainability_preference_profile_context,
 )
-from src.api.services.construction_execution_source_context import (
-    external_order_execution_acknowledgement_context,
-)
 from src.api.services.construction_liquidity_source_context import (
     client_income_needs_schedule_context,
     liquidity_cashflow_projection_context,
@@ -781,23 +778,3 @@ def test_external_treasury_currency_overlay_context_absent_without_source_respon
         )
         is None
     )
-
-
-def test_external_order_acknowledgement_context_is_fail_closed_source_evidence() -> None:
-    context = external_order_execution_acknowledgement_context(_acknowledgement_response())
-
-    assert context is not None
-    assert context.supportability_status == ConstructionMethodStatus.BLOCKED
-    assert context.source_system == "lotus-core"
-    assert context.source_product_name == "ExternalOrderExecutionAcknowledgement"
-    assert context.source_id == "core-ack-fingerprint"
-    assert context.acknowledgement_count == 0
-    assert context.blocked_capabilities == ["execution", "fill", "settlement"]
-    assert context.reason_codes == [
-        "EXTERNAL_OMS_SOURCE_NOT_INGESTED",
-        "EXTERNAL_ORDER_EXECUTION_ACKNOWLEDGEMENT_FAIL_CLOSED",
-    ]
-
-
-def test_external_order_acknowledgement_context_absent_without_source_response() -> None:
-    assert external_order_execution_acknowledgement_context(None) is None
